cilium-cli: move isSupersededPodRejection to the k8s package
The connectivity test setup needs the same notion of a pod that reached its
terminal state without ever running, so move the helper out of the status
package into cilium-cli/k8s, which both call paths already import, and export
it. The list of reasons, its behaviour and its test are unchanged, this only
moves them.

AIL:3
